reorganize interactions
