+++ prolif/interactions/__init__.py
@@ -1,0 +1,2 @@
+from prolif.interactions.base import Distance, DoubleAngle, Interaction, SingleAngle
+from prolif.interactions.interactions import *
